+++ ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: abel-haj <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/13 16:02:31 by abel-haj          #+#    #+#             */
-/*   Updated: 2019/10/18 03:16:03 by abel-haj         ###   ########.fr       */
+/*   Updated: 2019/11/01 17:46:20 by abel-haj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 void	*ft_calloc(size_t count, size_t size)
 {
-	unsigned int	i;
-	char			*ptr;
+	size_t	i;
+	char	*ptr;
 
 	i = 0;
 	ptr = (char *)malloc(count * size);
