+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: abel-haj <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/10 22:10:22 by abel-haj          #+#    #+#             */
-/*   Updated: 2019/10/18 23:39:43 by abel-haj         ###   ########.fr       */
+/*   Updated: 2019/11/01 23:27:29 by abel-haj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	char unsigned	*string;
 
 	i = 0;
