+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: abel-haj <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/11 19:12:09 by abel-haj          #+#    #+#             */
-/*   Updated: 2019/10/23 16:07:04 by abel-haj         ###   ########.fr       */
+/*   Updated: 2019/11/01 23:25:14 by abel-haj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*f_string;
 	unsigned char	*s_string;
