+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: abel-haj <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/11 19:02:15 by abel-haj          #+#    #+#             */
-/*   Updated: 2019/10/25 02:06:21 by abel-haj         ###   ########.fr       */
+/*   Updated: 2019/11/01 23:24:40 by abel-haj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,10 +15,10 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	unsigned long	i;
-	unsigned long	j;
-	unsigned long	needle_len;
-	char			*string;
+	size_t	i;
+	size_t	j;
+	size_t	needle_len;
+	char	*string;
 
 	i = 0;
 	j = i;
