+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: abel-haj <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/13 18:33:45 by abel-haj          #+#    #+#             */
-/*   Updated: 2019/10/25 08:07:40 by abel-haj         ###   ########.fr       */
+/*   Updated: 2019/11/01 23:34:15 by abel-haj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	unsigned int	i;
-	char			*string;
+	size_t	i;
+	char	*string;
 
 	if (!s)
 		return (NULL);
